DPO-1537 convert to maven, lots of changes

+++ src/main/java/fi/digitraffic/graphql/rail/to/CauseTOConverter.java
@@ -11,10 +11,10 @@ public class CauseTOConverter extends BaseConverter<CauseTO> {
     @Override
     public CauseTO convert(Tuple tuple) {
         return new CauseTO(
-                tuple.get(QCause.cause.timeTableRowId.attapId).intValue(),
-                tuple.get(QCause.cause.timeTableRowId.trainNumber).intValue(),
+                tuple.get(QCause.cause.timeTableRowId.attapId).longValue(),
+                tuple.get(QCause.cause.timeTableRowId.trainNumber).longValue(),
                 tuple.get(QCause.cause.timeTableRowId.departureDate),
-                tuple.get(QCause.cause.id).intValue(),
+                tuple.get(QCause.cause.id).longValue(),
                 tuple.get(QCause.cause.categoryCodeOid),
                 tuple.get(QCause.cause.detailedCategoryCodeOid),
                 tuple.get(QCause.cause.thirdCategoryCodeOid),
