+++ src/main/java/fi/digitraffic/graphql/rail/to/CompositionTOConverter.java
@@ -11,10 +11,10 @@ public class CompositionTOConverter {
     public CompositionTO convert(Tuple tuple) {
         return new CompositionTO(
                 tuple.get(QComposition.composition.id.departureDate),
-                tuple.get(QComposition.composition.id.trainNumber).intValue(),
+                tuple.get(QComposition.composition.id.trainNumber),
                 tuple.get(QComposition.composition.operatorShortCode),
-                tuple.get(QComposition.composition.trainCategoryId).intValue(),
-                tuple.get(QComposition.composition.trainTypeId).intValue(),
+                tuple.get(QComposition.composition.trainCategoryId),
+                tuple.get(QComposition.composition.trainTypeId),
                 tuple.get(QComposition.composition.version).toString(),
                 tuple.get(QComposition.composition.operatorUicCode),
                 null,
